% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_market.R
\name{gp}
\alias{gp}
\title{gp}
\usage{
gp(f, p)
}
\arguments{
\item{f}{the fit}

\item{p}{the parameter name}
}
\description{
get parameter point estimates from a Stan fit (optimizing, vb or sampling)
}
