% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrcl-package.R
\docType{package}
\name{rrcl-package}
\alias{rrcl-package}
\alias{rrcl}
\title{The 'rrcl' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
