% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_market.R
\name{sim_markets}
\alias{sim_markets}
\title{sim_markets}
\usage{
sim_markets(P, P2, J, T, NS = 500, alpha = 0.5, alpha_negative = 1,
  beta_mean = 1, sigma_beta = 1, sigma_tau = 0.5,
  gamma_intercept = 4, sigma_gamma = 0.3, market_size = 50000,
  xi_halfrange = 3)
}
\arguments{
\item{P}{number of choice attributes (these will be binary)}

\item{P2}{number of cost-shifting excluded instruments}

\item{J}{number of products to be simulated-- each market will have different numbers of product options}

\item{T}{number of markets}

\item{NS}{the number of fake decision-makers to simulate. 500 by default}

\item{alpha}{alpha_i ~ normal(alpha, ...) for each decision-maker, with alpha_i either being the
coefficient on price for a given decisionmaker (when alpha_negative = 0) or -exp(alpha) being their
coefficient on price. 0.5 by default}

\item{alpha_negative}{1 for alpha to be transformed to -exp(alpha), the default, 0 otherwise.}

\item{beta_mean}{the beta--"part worth" coefficients are drawn from beta ~ normal(beta_mean, sigma_beta)}

\item{sigma_beta}{see beta_mean}

\item{sigma_tau}{the marginal scale of the covariance matrix. tau ~ normal+(0, sigma_tau)}

\item{gamma_intercept}{the intercept in the marginal cost model. Increase this if you're getting negative
marginal costs. 4 by default.}

\item{sigma_gamma}{scale of the coefficients on the cost-shifters}

\item{market_size}{the number of participants in each market. 50k by default}

\item{xi_halfrange}{in the generative model, we assume xi is uniformly
distributed with zero mean. xi_halfrange is half the range (that is,
xi is uniform between -xi_halfrange, xi_halfrange). 3 by default.}
}
\value{
a data frame containing shares
}
\description{
Simulates market equilibrium prices for a number of markets given key structural parameters
under random coefficients logit and the assumption that each firm produces one product and
chooses only price to profit maximize.
}
