% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_arcl.R
\name{stan_arcl}
\alias{stan_arcl}
\title{stan_arcl}
\usage{
stan_arcl(data, NS = 100, sales = "sales", market_ids = "market_ids",
  prices = "prices", X_names, W_names, alpha_negative = 1, init = 0)
}
\arguments{
\item{data}{your dataset}

\item{NS}{the number of individuals to be simulated in each market}

\item{sales}{string. The column name in data where sales data belong. \code{sales} by default}

\item{market_ids}{string. Column name for the identifier of the markets.  \code{market} by default}

\item{prices}{string. Column name for the price.  \code{price} by default}

\item{X_names}{string. Column names for the product attributes}

\item{W_names}{string. Column names for excluded instruments that affect marginal costs}

\item{alpha_negative}{should we constrain alpha to be negative? 1 for negatgive, 0 otherwise.}

\item{init}{passed to optimizing}
}
\description{
Fits an aggregate random coefficients logit model in Stan
}
